feat: extractMarkdownHeadline
Former-commit-id: 4eb7df2ac38edc66f597daa1e0a41f1edf21a478

+++ backend/src/main/kotlin/com/storyteller_f/index/LuceneTopicSearchService.kt
@@ -52,7 +52,7 @@ class LuceneTopicSearchService(private val path: Path) : TopicSearchService {
                 DirectoryReader.open(it).use { reader ->
                     val searcher = IndexSearcher(reader)
                     idList.map { id ->
-                        val topDocs = searcher.search(LongPoint.newExactQuery("id1", id.toLong()), 1)
+                        val topDocs = searcher.search(LongPoint.newExactQuery("id1", id), 1)
                         topDocs.scoreDocs.firstOrNull()?.let { scoreDoc ->
                             val document = searcher.storedFields().document(scoreDoc.doc)
                             restoreDocument(id, document)
